@@ -3,19 +3,18 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@lingui/core": "^3.13.2",
-    "@lingui/react": "^3.13.2",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "@lingui/core": "^4.0.0-next.3",
+    "@lingui/react": "^4.0.0-next.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "scripts": {
-    "start": "yarn extract && yarn compile && react-scripts start",
-    "build": "yarn extract && yarn compile && react-scripts build",
+    "start": "yarn extract && react-scripts start",
+    "build": "yarn extract && react-scripts build",
     "test": "react-scripts test",
-    "test:ci": "yarn extract && yarn compile && react-scripts test --watchAll=false",
+    "test:ci": "yarn extract && react-scripts test --watchAll=false",
     "eject": "react-scripts eject",
-    "extract": "lingui extract",
-    "compile": "lingui compile"
+    "extract": "lingui extract"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -33,18 +32,17 @@
     ]
   },
   "devDependencies": {
-    "@lingui/cli": "^3.13.2",
-    "@lingui/loader": "^3.13.2",
-    "@lingui/macro": "^3.13.2",
-    "@testing-library/jest-dom": "^5.11.5",
-    "@testing-library/react": "^11.1.0",
-    "@testing-library/user-event": "^12.1.10",
-    "@types/jest": "^26.0.15",
-    "@types/node": "^14.14.6",
-    "@types/react": "^16.9.55",
-    "@types/react-dom": "^16.9.9",
-    "react-scripts": "4.0.0",
-    "typescript": "^4.0.5"
-  },
-  "packageManager": "yarn@1.22.19"
+    "@lingui/cli": "^4.0.0-next.3",
+    "@lingui/loader": "^4.0.0-next.3",
+    "@lingui/macro": "^4.0.0-next.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^29.5.0",
+    "@types/node": "^16.18.18",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.9.5"
+  }
 }
