@@ -1,12 +1,12 @@
-import React, { useEffect } from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import React, { useEffect } from "react"
+import "./index.css"
+import App from "./App"
+import * as serviceWorker from "./serviceWorker"
 
-import { i18n } from "@lingui/core";
-import { I18nProvider } from '@lingui/react'
-import { defaultLocale, dynamicActivate } from './i18n';
+import { i18n } from "@lingui/core"
+import { I18nProvider } from "@lingui/react"
+import { defaultLocale, dynamicActivate } from "./i18n"
+import * as ReactDOMClient from "react-dom/client"
 
 const I18nApp = () => {
   useEffect(() => {
@@ -16,19 +16,18 @@ const I18nApp = () => {
 
   return (
     <I18nProvider i18n={i18n}>
-      <App  />
+      <App />
     </I18nProvider>
   )
 }
 
-ReactDOM.render(
+ReactDOMClient.createRoot(document.getElementById("root") as Element).render(
   <React.StrictMode>
     <I18nApp />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+  </React.StrictMode>
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+serviceWorker.unregister()
